修复主机部署 DK 没有采集容器 limit 的问题

See merge request cloudcare-tools/datakit!3523

+++ internal/container/runtime/ecsfargate.go
@@ -116,7 +116,7 @@ func (c *ecsfargateClient) ContainerTop(id string) (*ContainerTop, error) {
 	// cpu usage
 	top.CPUPercent, top.CPUUsageMillicores = calculateCPUUsageUnix(&stats)
 	// cpu cores
-	top.CPUCores = int(stats.CPUStats.OnlineCPUs)
+	top.CPUCores = int64(stats.CPUStats.OnlineCPUs)
 
 	// memory usage and menory limit
 	top.MemoryWorkingSet = calculateMemUsageUnixNoCache(stats.MemoryStats)
